fix(E03-S14): corrige mais 8 achados reais do pgTAP na 2ª rodada de CI
Segunda leva de achados genuínos, agora que os fixtures de created_by/role
da 1ª rodada destravaram execução mais funda:

- comercial_leads_aposentado.test.sql: insert em auth.users acontecia
  DEPOIS de já trocar pro role authenticated (sem INSERT em auth.users) —
  movido pro topo, antes da troca de role.
- comercial_dashboard_rls.test.sql: is() comparava bigint (retorno de
  fn_ciclo_venda.quantidade) contra literal integer sem cast — 3 ocorrências,
  todas com ::int.
- comercial_contratos_rls.test.sql: insert direto em financeiro.contratos
  exige user_modulos.financeiro=escrita (RLS própria do Financeiro,
  migration 0108) — comercial:escrita não basta. 3 asserções que fazem join
  com financeiro.contratos após fn_ativar_contrato também precisavam de
  financeiro:leitura nas claims do caller (a RPC é security definer e cria
  o plano sem depender disso, mas a LEITURA do resultado pelo teste sim).
- comercial_fundacao_rls.test.sql: trigger de motivo de perda obrigatório
  lança 23514 (check_violation), não P0001 — assunção de código errada.
- comercial_levantamento_rls.test.sql: SELECT direto em pcm.inspecoes pra
  confirmar persistência rodava com a RLS do caller (comercial:escrita,
  sem pcm) — pcm.inspecoes exige módulo pcm pra leitura. Trocado pra
  service_role só nesse check, com role restaurado logo depois pro teste
  seguinte (que precisa da negação normal).
- comercial_portal_proposta_rls.test.sql: 2 blocos de asserção liam
  comercial.oportunidades/oportunidade_eventos direto como cliente-sindico
  — RLS dessas tabelas não tem carve-out pra síndico (só a view do portal
  tem). Trocado pra service_role nesses checks específicos, restaurando o
  role/claims de síndico antes da próxima chamada de RPC (cuja guarda
  exige o caller ser de fato cliente-sindico).

+++ supabase/tests/comercial_dashboard_rls.test.sql
@@ -35,7 +35,7 @@ select is(
   'sem modulo comercial: fn_conversao_etapas nao levanta erro, so devolve vazio (RLS filtra)'
 );
 select is(
-  (select quantidade from comercial.fn_ciclo_venda(current_date - 30, current_date)),
+  (select quantidade::int from comercial.fn_ciclo_venda(current_date - 30, current_date)),
   0,
   'sem modulo comercial: fn_ciclo_venda devolve quantidade zero'
 );
@@ -88,12 +88,12 @@ update comercial.oportunidades set etapa_id = '00000000-0000-0000-0000-000000000
 update comercial.oportunidades set fechada_em = now() - interval '5 days' where id = '00000000-0000-0000-0000-0000000009d9';
 
 select is(
-  (select quantidade from comercial.fn_ciclo_venda(current_date - 10, current_date)),
+  (select quantidade::int from comercial.fn_ciclo_venda(current_date - 10, current_date)),
   1,
   'ciclo de venda conta a oportunidade pelo fechamento MAIS RECENTE (5 dias atras), nao pelo primeiro (30 dias atras, fora desta janela)'
 );
 select is(
-  (select quantidade from comercial.fn_ciclo_venda(current_date - 35, current_date - 20)),
+  (select quantidade::int from comercial.fn_ciclo_venda(current_date - 35, current_date - 20)),
   0,
   'ciclo de venda NAO conta a oportunidade na janela do PRIMEIRO fechamento (ja superado pela reabertura)'
 );
